-Code refactoring -moved some unreferenced code to file marked to be not compiled -marked several camera functions static for performance
