+++ LitePlacer/packages.config
@@ -7,5 +7,10 @@
   <package id="AForge.Video" version="2.2.5" targetFramework="net40" />
   <package id="AForge.Video.DirectShow" version="2.2.5" targetFramework="net40" />
   <package id="MathNet.Numerics" version="4.8.0" targetFramework="net40" />
+  <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.9.6" targetFramework="net40" developmentDependency="true" />
+  <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="2.9.6" targetFramework="net40" developmentDependency="true" />
+  <package id="Microsoft.CodeQuality.Analyzers" version="2.9.6" targetFramework="net40" developmentDependency="true" />
+  <package id="Microsoft.NetCore.Analyzers" version="2.9.6" targetFramework="net40" developmentDependency="true" />
+  <package id="Microsoft.NetFramework.Analyzers" version="2.9.6" targetFramework="net40" developmentDependency="true" />
   <package id="Newtonsoft.Json" version="12.0.2" targetFramework="net40" />
 </packages>
